Final_Analysis/requirements.txt
pika==1.3.2
numpy==2.0.0
pandas==2.2.2
uproot==5.3.9
matplotlib==3.9.0
lmfit==1.3.1
awkward-pandas==2023.8.0
aiohttp==3.9.5
requests==2.32.3
vector
matplotlib
lmfit